producer/requirements.txt
dependency-injector==4.41.0
fastapi==0.94.1
mysqlclient==2.2.0
kafka-python==2.0.2
pydantic==1.9.1
pytest==7.3.1
pytest-env==0.6.2
PyYAML==6.0
SQLAlchemy==2.0.22
uvicorn==0.20.0a
httpx==0.25.0